refactor: move DB logic for lesson list from service to repository
