+++ src/lesson/lesson.repository.ts
@@ -5,8 +5,10 @@ import {
   LessonRequest,
   LessonRequestStatus,
   Prisma,
+  Region,
 } from '@prisma/client';
 import { PrismaService } from '#prisma/prisma.service.js';
+import { QueryLessonDto } from './dto/lesson.dto.js';
 import { ILessonRepository } from './interface/lesson-repository.interface.js';
 import { CreateLesson, LessonResponse, PatchLesson } from './type/lesson.type.js';
 
@@ -18,6 +20,35 @@ export class LessonRepository implements ILessonRepository {
     this.lessonRequest = prisma.lessonRequest;
     this.directQuoteRequest = prisma.directQuoteRequest;
   }
+
+  private regionMapping: Record<Region, string[]> = {
+    SEOUL: ['서울'],
+    GYEONGGI: ['경기'],
+    INCHEON: ['인천'],
+    GANGWON: ['강원', '강원특별자치도'],
+    CHUNGBUK: ['충북'],
+    CHUNGNAM: ['충남'],
+    JEONBUK: ['전북', '전북특별자치도'],
+    JEONNAM: ['전남'],
+    GYEONGBUK: ['경북'],
+    GYEONGNAM: ['경남'],
+    DAEGU: ['대구'],
+    DAEJEON: ['대전'],
+    BUSAN: ['부산'],
+    ULSAN: ['울산'],
+    GWANGJU: ['광주'],
+    JEJU: ['제주', '제주특별자치도'],
+  };
+
+  private orderMapping: Record<string, string> = {
+    created_at: 'createdAt',
+    start_date: 'startDate',
+    end_date: 'endDate',
+    quote_end_date: 'quoteEndDate',
+    rating: 'rating',
+    lesson_count: 'lessonCount',
+    lesson_time: 'lessonTime',
+  };
 
   async create(data: CreateLesson & { userId: string }): Promise<LessonRequest> {
     return await this.lessonRequest.create({ data });
@@ -44,6 +75,187 @@ export class LessonRepository implements ILessonRepository {
     return await this.lessonRequest.count({
       where,
     });
+  }
+
+  async findLessons(
+    query: QueryLessonDto,
+    currentUserId: string,
+    myLessonUserId?: string,
+  ): Promise<{ lessons: LessonResponse[]; totalCount: number }> {
+    const {
+      page = 1,
+      limit = 5,
+      order = 'createdAt',
+      sort = 'desc',
+      keyword,
+      lessonType,
+      lessonSubType,
+      locationType,
+      status,
+      gender,
+      region,
+      hasDirectQuote,
+    } = query.toCamelCase();
+
+    // 필터(where) 조건 구성
+    const where: Prisma.LessonRequestWhereInput = {
+      ...(myLessonUserId && { userId: myLessonUserId }),
+      ...(lessonType && { lessonType: { in: lessonType } }),
+      ...(lessonSubType && { lessonSubType: { in: lessonSubType } }),
+      ...(locationType && { locationType: { in: locationType } }),
+      ...(status && { status: { in: status } }),
+      ...(gender && {
+        user: {
+          is: {
+            profile: {
+              is: {
+                gender: { in: gender },
+              },
+            },
+          },
+        },
+      }),
+      ...(hasDirectQuote !== undefined && {
+        directQuoteRequests: hasDirectQuote
+          ? { some: { trainerId: currentUserId } }
+          : { none: { trainerId: currentUserId } },
+      }),
+      AND: [
+        // region 필터
+        ...(region
+          ? [
+              {
+                OR: region.flatMap(
+                  (r) =>
+                    this.regionMapping[r]?.map((koreanRegion) => ({
+                      roadAddress: { contains: koreanRegion },
+                    })) || [],
+                ),
+              },
+            ]
+          : []),
+
+        ...(keyword
+          ? [
+              {
+                OR: [
+                  // nickname 검색
+                  {
+                    user: {
+                      is: {
+                        nickname: {
+                          contains: keyword,
+                          mode: Prisma.QueryMode.insensitive,
+                        },
+                      },
+                    },
+                  },
+                  // profile.name 검색
+                  {
+                    user: {
+                      is: {
+                        profile: {
+                          is: {
+                            name: {
+                              contains: keyword,
+                              mode: Prisma.QueryMode.insensitive,
+                            },
+                          },
+                        },
+                      },
+                    },
+                  },
+                ],
+              },
+            ]
+          : []),
+      ],
+    };
+
+    const orderByField = this.orderMapping[order] || 'createdAt';
+    const orderBy: Record<string, any> = {};
+    orderBy[orderByField] = sort;
+
+    const skip = (page - 1) * limit;
+    const take = limit;
+
+    const select: Prisma.LessonRequestSelect = {
+      id: true,
+      userId: true,
+      lessonType: true,
+      lessonSubType: true,
+      startDate: true,
+      endDate: true,
+      lessonCount: true,
+      lessonTime: true,
+      quoteEndDate: true,
+      locationType: true,
+      postcode: true,
+      roadAddress: true,
+      detailAddress: true,
+      status: true,
+      createdAt: true,
+      updatedAt: true,
+      directQuoteRequests: {
+        select: {
+          id: true,
+          lessonRequestId: true,
+          trainerId: true,
+          status: true,
+          rejectionReason: true,
+        },
+      },
+      lessonQuotes: {
+        select: {
+          id: true,
+          lessonRequestId: true,
+          trainerId: true,
+          price: true,
+          status: true,
+          createdAt: true,
+          updatedAt: true,
+          rejectionReason: true,
+          trainer: {
+            select: {
+              id: true,
+              nickname: true,
+              profile: {
+                select: {
+                  name: true,
+                  region: true,
+                },
+              },
+            },
+          },
+        },
+      },
+      user: {
+        select: {
+          id: true,
+          nickname: true,
+          profile: {
+            select: {
+              name: true,
+              gender: true,
+              region: true,
+            },
+          },
+        },
+      },
+    };
+
+    const [lessons, totalCount] = await Promise.all([
+      this.lessonRequest.findMany({
+        where,
+        orderBy,
+        skip,
+        take,
+        select,
+      }),
+      this.lessonRequest.count({ where }),
+    ]);
+
+    return { lessons, totalCount };
   }
 
   async findLessonsByUserId(userId: string, status?: LessonRequestStatus): Promise<LessonRequest[]> {
@@ -126,7 +338,7 @@ export class LessonRepository implements ILessonRepository {
     }));
   }
 
-  // 남/여 카운트 계산을 위한 조회회
+  // 남/여 카운트 계산을 위한 조회
   async findAllForGenderCount() {
     return this.prisma.lessonRequest.findMany({
       select: {
