+++ src/lesson/type/lesson.type.ts
@@ -5,6 +5,7 @@ import {
   LessonType,
   LocationType,
   Profile,
+  Role,
 } from '@prisma/client';
 
 export type CreateLesson = {
@@ -24,7 +25,13 @@ export type CreateLesson = {
 export type PatchLesson = Partial<CreateLesson>;
 
 export type LessonResponse = LessonRequest & {
+  // user?: {
+  //   profile?: Profile;
+  // };
   user?: {
+    // Prisma에서 오는 필드 전부
+    id: string;
+    nickname: string;
     profile?: Profile;
   };
   directQuoteRequests?: {
